+++ vscode-live-preview/src/extension.ts
@@ -1,35 +1,28 @@
 import * as vscode from "vscode";
-import { InlinePanel } from "./panel-controller/inlinePanel.js";
 import statusBar from "./utils/status-bar.js";
+import ConsoleOutput from "./utils/output-channel.js";
+import { configChangeHandler, userConfig, userCustom } from "./utils/config.js";
+import { InlinePanel } from "./panel-controller/inlinePanel.js";
 import { openBrowser } from "./utils/browser.js";
 import { findHTMLDocument } from "./utils/htmldoc.js";
-import { userConfig, userCustom } from "./utils/config.js";
-import ConsoleOutput from "./utils/output-channel.js";
 import { launchDebug } from "./utils/debug.js";
-import { Command } from "./utils/constant.js";
+import { Command, LaunchBrowsers } from "./utils/constant.js";
+import { showQuickPick } from "./utils/quick-pick.js";
 
 const workspaceFolder = vscode.workspace.workspaceFolders[0].uri.fsPath;
 
 export async function activate(context: vscode.ExtensionContext) {
 	let port = userConfig.serverPort || 2200;
 	let previewServer;
-	const disposableBrowser = vscode.commands.registerCommand(Command.LaunchDefault, async () => {
-		const docPath = await startPreviewServer();
-		openBrowser("default", `${port}${docPath}`);
-		statusBar.setCloseCommand();
-	});
 
-	const disposableChrome = vscode.commands.registerCommand(Command.LaunchChrome, async () => {
-		const docPath = await startPreviewServer();
-		openBrowser("chrome", `${port}${docPath}`);
-		statusBar.setCloseCommand();
-	});
-
-	const disposableFirefox = vscode.commands.registerCommand(Command.LaunchFirefox, async () => {
-		const docPath = await startPreviewServer();
-		openBrowser("firefox", `${port}${docPath}`);
-		statusBar.setCloseCommand();
-	});
+	for (const browser in LaunchBrowsers) {
+		const disposableBrowser = vscode.commands.registerCommand(LaunchBrowsers[browser], async () => {
+			const docPath = await startPreviewServer();
+			openBrowser(browser, `${port}${docPath}`);
+			statusBar.setCloseCommand();
+		});
+		context.subscriptions.push(disposableBrowser);
+	}
 
 	const disposablePanel = vscode.commands.registerCommand(Command.InlinePreview, async () => {
 		const docPath = await startPreviewServer();
@@ -47,29 +40,32 @@ export async function activate(context: vscode.ExtensionContext) {
 		statusBar.setCloseCommand();
 	});
 
+	const disposableQuickpick = vscode.commands.registerCommand(Command.PickBrowser, showQuickPick);
+
 	//close server on command
 	const disposableClose = vscode.commands.registerCommand(Command.CloseServer, () => {
 		closePreview();
 		statusBar.setStartCommand();
 	});
+	const disposableOnConfig = vscode.workspace.onDidChangeConfiguration(configChangeHandler);
 
-	context.subscriptions.push(disposableBrowser);
-	context.subscriptions.push(disposableChrome);
-	context.subscriptions.push(disposableFirefox);
 	context.subscriptions.push(disposableServer);
+	context.subscriptions.push(disposableQuickpick);
 	context.subscriptions.push(disposableDebug);
 	context.subscriptions.push(disposablePanel);
 	context.subscriptions.push(disposableClose);
+	context.subscriptions.push(disposableOnConfig);
 
 	async function startPreviewServer() {
 		const { PreviewServer } = require("../../preview-server/build/server.js");
-		// const { PreviewServer } = await import("../../preview-server/server.js");
+		//const { PreviewServer } = await import("../../preview-server/server.js");
 		const serverLogger = new ConsoleOutput("LPS server log");
 		const extPath = context.extensionPath;
 		//start preview server
 		previewServer = new PreviewServer(workspaceFolder, extPath, userConfig, serverLogger, userCustom);
 		port = await previewServer.startServer(port).catch(async (err) => console.error(err));
 		if (!port) return vscode.window.showErrorMessage("Cannot start server");
+
 		//find html file in current directory
 		const document = await findHTMLDocument().catch((err) => console.error(err));
 		await new Promise((r) => setTimeout(r, 100));
@@ -104,26 +100,23 @@ export async function activate(context: vscode.ExtensionContext) {
 			previewServer.changePageUrl(textDocument);
 		});
 
-		/* const disposableOnOpen = vscode.workspace.onDidOpenTextDocument(async (textDocument) => {
-			previewServer.onTxtDocumentActive(textDocument);
-		});
-
-		context.subscriptions.push(disposableOnOpen); */
 		context.subscriptions.push(disposableOnDidSave);
 		context.subscriptions.push(disposableOnDidChange);
 
 		if (userConfig.liveRefresh) {
 			var disposableOnChange = vscode.workspace.onDidChangeTextDocument((textDocumentChangeEvent) => {
 				const { document, contentChanges } = textDocumentChangeEvent;
-				if (!contentChanges[0]) return;
-				//TODO for-loop  on contentChanges
-				switch (document.languageId) {
-					case "html":
-						previewServer.updateElementAtPosition(contentChanges[0], document);
-						break;
-					case "css":
-						previewServer.updateRuleAtPosition(contentChanges[0], document);
-						break;
+				if (!Array.isArray(contentChanges)) return;
+
+				for (const contentChange of contentChanges) {
+					switch (document.languageId) {
+						case "html":
+							previewServer.updateElementAtPosition(contentChange, document);
+							break;
+						case "css":
+							previewServer.updateRuleAtPosition(contentChange, document);
+							break;
+					}
 				}
 			});