+++ vscode-live-preview/src/utils/browser.ts
@@ -7,21 +7,34 @@ const defaultBrowser = {
 };
 
 const win32 = {
-	chrome: "start chrome",
-	firefox: "start firefox",
-	brave: "start brave",
+	Brave: "start brave",
+	Chrome: "start chrome",
+	Chromium: "start chromium-browser",
+	Edge: "start msedge",
+	Firefox: "start firefox",
+	Opera: "start opera",
+	Vivaldi: "start vivaldi",
 };
 
 const linux = {
-	chrome: "google-chrome",
-	firefox: "firefox",
-	brave: "brave-browser",
+	Brave: "brave-browser",
+	Chrome: "google-chrome",
+	Chromium: "chromium-browser",
+	Edge: "msedge",
+	Firefox: "firefox",
+	Opera: "opera",
+	Vivaldi: "vivaldi",
 };
 
 const darwin = {
-	chrome: 'open -a "Google Chrome" -g ',
-	firefox: "open -a firefox -g ",
-	brave: 'open -a "Brave Browser"',
+	Brave: 'open -a "Brave Browser"',
+	Chrome: 'open -a "Google Chrome" -g ',
+	Chromium: "open -a chromium",
+	Edge: 'open -a  "Microsoft Edge"',
+	Firefox: "open -a firefox -g ",
+	Safari: 'open -a "Brave Browser"',
+	Opera: "open -a Opera",
+	Vivaldi: "open -a vivaldi",
 };
 
 const platforms = {
@@ -30,11 +43,11 @@ const platforms = {
 	darwin: darwin,
 };
 
-function getBrowserCommand(browser, platform, url) {
+function getBrowserCommand(browser: string, platform: string, url: string) {
 	url = "http://localhost:" + url;
-	if (browser === "default") return `${defaultBrowser[platform]} ${url}`;
+	if (browser === "Default") return `${defaultBrowser[platform]} ${url}`;
 	let browserCmd = `${platforms[platform][browser]} ${url}`;
-	platform === "linux" && (browserCmd += " &");
+	platform === "linux" && (browserCmd += " & disown");
 	return browserCmd;
 }
 
@@ -42,5 +55,5 @@ export function openBrowser(browser = "default", url: string) {
 	const terminal = vscode.window.createTerminal(`live preview server`);
 	const browserCmd = getBrowserCommand(browser, process.platform, url);
 	terminal.sendText(browserCmd);
-	setTimeout(() => terminal.dispose(), 1000);
+	setTimeout(() => terminal.dispose(), 2000);
 }