+++ vscode-live-preview/src/utils/config.ts
@@ -6,6 +6,7 @@ type UserConfig = {
 	liveRefresh: boolean;
 	customElementHMR: boolean;
 	compileTs: boolean;
+	statusBarCommand: string;
 	defaultBrowser: string;
 	serverPort: number;
 	baseDir: string;
@@ -15,12 +16,14 @@ type UserConfig = {
 
 type UserCustom = {
 	httpHeaders: object;
+	quickPickBrowsers: string[];
 };
 
 export const userConfig: UserConfig = {
 	liveRefresh: configMap.get("liveRefresh"),
-	customElementHMR: configMap.get("customElementHMRectory"),
-	defaultBrowser: configMap.get("defaultBrowserectory"),
+	customElementHMR: configMap.get("customElementHMR"),
+	defaultBrowser: configMap.get("defaultBrowser"),
+	statusBarCommand: configMap.get("statusBarCommand"),
 	serverPort: configMap.get("serverPort"),
 	baseDir: configMap.get("baseDir"),
 	compileTs: configMap.get("compileTs"),
@@ -30,4 +33,18 @@ export const userConfig: UserConfig = {
 
 export const userCustom: UserCustom = {
 	httpHeaders: customMap.get("httpHeaders"),
+	quickPickBrowsers: customMap.get("quickPickBrowsers"),
 };
+
+export function configChangeHandler({ affectsConfiguration }) {
+	//config check
+	const isAffected = affectsConfiguration("livePreviewServer.config");
+	if (isAffected) {
+		for (const configKey in userConfig) userConfig[configKey] = configMap.get(configKey);
+	}
+	//custom check
+	const isAffected2 = affectsConfiguration("livePreviewServer.custom");
+	if (isAffected2) {
+		for (const customKey in userCustom) userCustom[customKey] = customMap.get(customKey);
+	}
+}