+++ vscode-live-preview/src/utils/constant.ts
@@ -1,16 +1,19 @@
 export const Command = {
 	InlinePreview: "livePreviewServer.start.inlinePreview",
-	LaunchChrome: "livePreviewServer.start.chromeBrowser",
-	LaunchBrave: "livePreviewServer.start.braveBrowser",
-	LaunchFirefox: "livePreviewServer.start.firefoxBrowser",
-	LaunchDefault: "livePreviewServer.start.defaultBrowser",
 	StartServer: "livePreviewServer.start.lpsServer",
 	DebugServer: "livePreviewServer.start.debugServer",
 	CloseServer: "livePreviewServer.close.server",
+	PickBrowser: "livePreviewServer.start.PickBrowser",
 };
 
 export const LaunchBrowsers = {
+	Default: "livePreviewServer.start.defaultBrowser",
 	Chrome: "livePreviewServer.start.chromeBrowser",
 	Firefox: "livePreviewServer.start.firefoxBrowser",
 	Brave: "livePreviewServer.start.braveBrowser",
+	Safari: "livePreviewServer.start.safariBrowser",
+	Chromium: "livePreviewServer.start.chromiumBrowser",
+	Edge: "livePreviewServer.start.edgeBrowser",
+	Vivaldi: "livePreviewServer.start.vivaldiBrowser",
+	Opera: "livePreviewServer.start.operaBrowser",
 };