+++ CHANGELOG.md
@@ -5,3 +5,17 @@
 - css and html live refresh fix
 
 - fix window path issue
+
+### [0.0.9] - 2023/06/20
+
+### Fixed
+
+- Browser open then close bug fixed
+- Only reload fetched files
+- Cann't restart server bug fixed
+
+### Added
+
+- Browser quickpick list added
+- Configuration now synced
+- Status bar style updated